moar refactor
