+++ agent/software/software.go
@@ -1,0 +1,3 @@
+package software
+
+func SendSoftware() {}
