+++ agent/utils.go
@@ -238,21 +238,6 @@ func Unzip(src, dest string) error {
 	return nil
 }
 
-// https://yourbasic.org/golang/formatting-byte-size-to-human-readable-format/
-func ByteCountSI(b uint64) string {
-	const unit = 1024
-	if b < unit {
-		return fmt.Sprintf("%d B", b)
-	}
-	div, exp := int64(unit), 0
-	for n := b / unit; n >= unit; n /= unit {
-		div *= unit
-		exp++
-	}
-	return fmt.Sprintf("%.1f %cB",
-		float64(b)/float64(div), "kMGTPE"[exp])
-}
-
 func randRange(min, max int) int {
 	rand.Seed(time.Now().UnixNano())
 	return rand.Intn(max-min) + min
